requirements.txt
django
django-taggit
django-crispy-forms
pillow
django-import-export
django-jazzmin

405 {% url 'mainApp:inbox_detail' m.reciever.username %}
417 {% url 'mainApp:inbox_detail' m.sender.username %}
432 {% url 'mainApp:inbox' %}
433 {% url 'mainApp:group_inbox' %}
508 {% url 'mainApp:feed' %}
543 {% url 'mainApp:games' %}
626 {% url 'mainApp:inbox_detail' f.username %}

442 {% url 'userauths:my-profile' %}
462 {% url 'userauths:profile-update' %}
458 {% url 'userauths:my-profile' %}
660 {% url 'userauths:sign-out' %}
637 {% url 'userauths:profile-update' %}
485 {% url 'userauths:sign-out' %}